Entity Catalog: Store: Move CF reducers (2)
